NXP-30044: extract S3TestHelper
